% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{All_Game_Attendance}
\alias{All_Game_Attendance}
\title{All Game Attendance Data}
\format{
A data frame with 6214 rows and 13 variables:
\describe{
\item{...1}{Number of games played}
\item{Year}{Integer. The year the teams played.}
\item{Day of Week}{Factor. The day of the week the teams played.}
\item{Month}{Factor. The month the teams played.}
\item{Day}{Integer. Day of the week the teams played.}
\item{Time(ET)}{The time the game was played at.}
\item{Game Type}{Factor. The game is classified either as regular season, Seminfinals, Finals, Eastern Conference Semifinals etc.}
\item{Home Team}{Fcator. Name of the home team.}
\item{Away Team}{Factor. Name of the away team.}
\item{Arena}{Factor. Name of the arena the game was played at.}
\item{City}{Factor. Name of the city the game was played at.}
\item{State}{Factor. Name of the State the game was played at.}
\item{Attendance}{Integer. Number of people who attended the game.}
}
}
\source{
\url{https://www.acrossthetimeline.com/wnba/data.html#}
}
\usage{
data(All_Game_Attendance)
}
\description{
This data set contains information about the attendance of all games in a certain season or series.
It includes columns such as game date, attendance numbers, and other related statistics.
}
\keyword{datasets}
