% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_ppg.R
\name{career_ppg}
\alias{career_ppg}
\title{Function shows career PPG of top WNBA scorers}
\source{
\url{https://www.basketball-reference.com/wnba/players/}
}
\usage{
career_ppg(players = NULL)
}
\arguments{
\item{players}{the first and last name of WNBA player(s)}
}
\value{
An interactive plot showing the career points scored by WNBA players over the years.
}
\description{
This function scrapes career points per game (PPG) data for specified WNBA players from Basketball Reference and returns an interactive plot showing points scored over the years.
}
\details{
This function fetches career points (PTS) and year (Year) data for top WNBA players, converts the data to numeric, and visualizes it using an interactive plot.
}
\examples{
career_ppg()
career_ppg("Diana Taurasi")
}
\author{
Jane Loo
}
