% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_WNBA_teams.R
\docType{data}
\name{globals_v}
\alias{globals_v}
\title{Function shows map of all WNBA teams past and present at any given time}
\format{
An object of class \code{character} of length 7.
}
\source{
\url{https://www.basketball-reference.com/wnba/teams/}
}
\usage{
globals_v
}
\value{
A US Map of all WNBA teams geographical location.
}
\description{
Function shows map of all WNBA teams past and present at any given time
}
\details{
Gives historical data of how WNBA teams have changed or moved over time.
}
\examples{
map_WNBA_teams(year_range = 2021:2024)
map_WNBA_teams(year_range = 1997:2024)
}
\author{
Jane Loo
}
\keyword{datasets}
