% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_WNBA_teams.R
\name{map_WNBA_teams}
\alias{map_WNBA_teams}
\title{Function shows map of all WNBA teams past and present at any given time}
\source{
\url{https://www.basketball-reference.com/wnba/teams/}
}
\usage{
map_WNBA_teams(year_range = 1997:2024)
}
\arguments{
\item{year_range}{a desired time period of years, formatted as "firstyear:lastyear"}
}
\value{
A US Map of all WNBA teams geographical location.
}
\description{
Function shows map of all WNBA teams past and present at any given time
}
\details{
Gives historical data of how WNBA teams have changed or moved over time.
}
\examples{
map_WNBA_teams(year_range = 2021:2024)
map_WNBA_teams(year_range = 1997:2024)
}
\author{
Jane Loo
}
