% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{plot_wnba_attendance}
\alias{plot_wnba_attendance}
\title{Function graphs a WNBA team attendance record over the years}
\source{
\url{https://www.acrossthetimeline.com/wnba/data.html#}
}
\usage{
plot_wnba_attendance(team)
}
\arguments{
\item{team}{string containing the name of the team}
}
\value{
graph
}
\description{
Function graphs a WNBA team attendance record over the years
}
\examples{
plot_wnba_attendance("Los Angeles Sparks")
}
\author{
Alina Abdulina
}
