% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winRate.R
\name{winRate}
\alias{winRate}
\title{Calculate the franchise history winning percentage of a WNBA team.}
\source{
\url{https://www.basketball-reference.com}
}
\usage{
winRate(team_name)
}
\arguments{
\item{team_name}{A string (either the city, the team name, or both) representing the name of the WNBA team.}
}
\value{
A numeric vector of length 1, representing the winning percentage of the team. Returns NA if no games are found.
}
\description{
Calculate the franchise history winning percentage of a WNBA team.
}
\examples{
winRate("Storm")
winRate("Seattle")
winRate("Seattle Storm")
}
\author{
Ella Sylvester
}
